customers/requirements.txt
uvicorn==0.18.2
fastapi==0.79.0
grpcio==1.47.0
grpcio-tools==1.47.0
supervisor==4.2.4
SQLAlchemy==1.4.39
databases[postgresql]==0.6.0
